feat: 주문 ui 핵심 뼈대 추가
