+++ clone/src/main/java/com/prgrammers/clone/config/MvcConfiguration.java
@@ -1,12 +1,14 @@
 package com.prgrammers.clone.config;
 
+import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+@Configuration
 public class MvcConfiguration implements WebMvcConfigurer {
 
 	@Override
 	public void addCorsMappings(CorsRegistry registry) {
-		registry.addMapping("/api/&&").allowedOrigins("*");
+		registry.addMapping("/api/**").allowedOrigins("*");
 	}
 }
