+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "axios": "^0.27.2",
     "bootstrap": "^5.1.3"
   }
 }
